+++ Engine/src/game_objects/SkinnedGameObject.cpp
@@ -1,6 +1,6 @@
 #include"SkinnedGameObject.h"
 
-SkinnedGameObject::SkinnedGameObject(unsigned int id, char* animationPath, Model* characterModel, glm::vec3 pos, glm::vec3 size, glm::vec3 rotation) : GameObject(id, characterModel, pos, size, rotation)
+SkinnedGameObject::SkinnedGameObject(uint8_t id, char* animationPath, Model* characterModel, glm::vec3 pos, glm::vec3 size, glm::vec3 rotation) : GameObject(id, characterModel, pos, size, rotation)
 {
 	mation = new Animation(animationPath, characterModel);
 	mator = new Animator(mation);
